Refactor helpers directory structure and update imports
